move styled components into their own files, move component files into a subdirectory
