+++ client/src/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App.jsx';
+import App from './components/App.jsx';
 
 ReactDOM.render(<App listingId={0} />, document.getElementById('photo-gallery'));
