+++ client/src/styled_components/scrollbar.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import styled from 'styled-components';
 
 var ScrollbarDiv = styled.div`
@@ -40,16 +39,4 @@ var Img = styled.img`
   transform: translateX(-50%) translateY(-50%);
 `;
 
-var Scrollbar = function({photos, currentPhotoIndex, viewPhotoHandler}) {
-  return (
-    <ScrollbarDiv>
-      {photos.map((photo, index) => (
-        <Thumbnail key={index} disabled={index === currentPhotoIndex} onClick={() => viewPhotoHandler(index)}>
-          <Img src={photo.url} />
-        </Thumbnail>
-      ))}
-    </ScrollbarDiv>
-  );
-};
-
-export default Scrollbar;
+export {ScrollbarDiv, Thumbnail, Img};
